+++ src/pages/not-found-404/index.ts
@@ -1,0 +1,1 @@
+export { NotFound404 } from './not-found-404';
