--- src/pages/not-fount-404/index.ts
@@ -1,1 +1,0 @@
-export { NotFound404 } from './not-fount-404';
